# swagger.yaml
openapi: 3.0.0
info:
  title: 'API Livros'
  description: 'Documentação da API Livros'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080/api-livros'
    description: 'Servidor local'
paths:
  /register:
    post:
      tags:
        - Auth
      summary: 'Realiza o cadastro de um novo usuário'
      operationId: 6ca1953c104d60936b62117784734f90
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 'Usuário registrado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: 'Dados inválidos ou email já cadastrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Erro interno ao cadastrar usuário'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      tags:
        - Auth
      summary: 'Realiza login do usuário'
      operationId: f6b87925c502a2f52803eb2945e0b2f6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 'Login realizado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: 'Erro de validação na requisição'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Credenciais inválidas'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /perfil:
    get:
      tags:
        - Auth
      summary: 'Retorna os dados do perfil do usuário autenticado'
      operationId: 11b2b4bdfe5b35bed6f42bed010cae98
      responses:
        '200':
          description: 'Perfil acessado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerfilResponse'
        '401':
          description: 'Token ausente, inválido ou sem permissão'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  /livros:
    get:
      tags:
        - Livros
      summary: 'Lista livros do usuário autenticado'
      operationId: c96a15ceb383a6a24fcdf8b8c54bef72
      parameters:
        -
          name: page
          in: query
          description: 'Página atual'
          required: false
          schema:
            type: integer
            example: 1
            default: 1
        -
          name: limit
          in: query
          description: 'Quantidade de itens por página'
          required: false
          schema:
            type: integer
            example: 10
            default: 10
      responses:
        '200':
          description: 'Livros retornados com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LivroListResponse'
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Nenhum livro encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Livros
      summary: 'Atualiza um livro existente do usuário autenticado'
      operationId: ee928fab664b23084de1dbe7464ea81e
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LivroUpdateRequest'
      responses:
        '200':
          description: 'Livro atualizado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LivroUpdateResponse'
        '400':
          description: 'Id do livro ausente ou dados inválidos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Livro não encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Erro ao atualizar livro'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Livros
      summary: 'Cria um novo livro para o usuário autenticado'
      operationId: b739368170d28460413bda43250d382b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LivroCreateRequest'
      responses:
        '201':
          description: 'Livro criado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LivroCreateResponse'
        '400':
          description: 'Dados inválidos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Não autenticado'
        '500':
          description: 'Erro ao criar livro'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Livros
      summary: 'Deleta um livro do usuário autenticado'
      operationId: d4487ace35a2539f559bc45a6a2a4568
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LivroDeleteRequest'
      responses:
        '204':
          description: 'Livro deletado com sucesso'
        '400':
          description: 'Id do livro não informado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Livro não encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    RegisterRequest:
      required:
        - nome
        - email
        - senha
      properties:
        nome:
          type: string
          example: 'Luiz Felipe'
          maxLength: 255
        email:
          type: string
          format: email
          example: teste@email.com
        senha:
          type: string
          format: password
          example: 12345678Ab@
      type: object
    LoginRequest:
      required:
        - email
        - senha
      properties:
        email:
          type: string
          format: email
          example: teste@email.com
        senha:
          type: string
          format: password
          example: 12345678Ab@
      type: object
    UsuarioPerfilResource:
      properties:
        id_usuario:
          type: integer
          example: 1
        UUID:
          type: string
          example: e04230085f34fcdc518137ac826725
        nome:
          type: string
          example: 'Luiz Felipe'
        email:
          type: string
          format: email
          example: teste@email.com
      type: object
    LoginDetailResource:
      properties:
        mensagem:
          type: string
          example: 'Login realizado com sucesso'
        access_token:
          type: string
          example: jwt_aqui
        UUID:
          type: string
          example: e04230085f34fcdc518137ac826725
        nome:
          type: string
          example: 'Luiz Felipe'
        email:
          type: string
          format: email
          example: teste@email.com
        foto_perfil:
          type: string
          example: null
          nullable: true
      type: object
    RegisterResponse:
      properties:
        success:
          type: boolean
          example: true
        detail:
          properties:
            mensagem:
              type: string
              example: 'Usuário registrado com sucesso'
          type: object
      type: object
    LoginResponse:
      properties:
        success:
          type: boolean
          example: true
        detail:
          $ref: '#/components/schemas/LoginDetailResource'
      type: object
    PerfilResponse:
      properties:
        success:
          type: boolean
          example: true
        detail:
          properties:
            mensagem:
              type: string
              example: 'Perfil acessado com sucesso'
            id_usuario:
              type: integer
              example: 1
            usuario:
              $ref: '#/components/schemas/UsuarioPerfilResource'
          type: object
      type: object
    ErrorResponse:
      properties:
        success:
          type: boolean
          example: false
        mensagem:
          type: string
          example: 'Erro ao processar requisição'
      type: object
    LivroResource:
      properties:
        id:
          type: integer
          example: 10
        titulo:
          type: string
          example: 'Dom Casmurro'
        autor:
          type: string
          example: 'Machado de Assis'
        ano:
          type: integer
          example: 1899
          nullable: true
        genero:
          type: string
          example: Romance
          nullable: true
        status:
          type: string
          example: quero_ler
        avaliacao:
          type: integer
          example: 5
          nullable: true
        anotacoes:
          type: string
          example: 'Clássico da literatura brasileira'
          nullable: true
      type: object
    LivroCreateRequest:
      required:
        - titulo
        - autor
        - ano
        - status
      properties:
        titulo:
          type: string
          example: 'Dom Casmurro'
        autor:
          type: string
          example: 'Machado de Assis'
        ano:
          type: integer
          example: 1899
        genero:
          type: string
          example: Romance
          nullable: true
        status:
          type: string
          example: quero_ler
        avaliacao:
          type: integer
          example: 5
          nullable: true
        anotacoes:
          type: string
          example: 'Clássico da literatura brasileira'
          nullable: true
      type: object
    LivroUpdateRequest:
      required:
        - id_livro
      properties:
        id_livro:
          type: integer
          example: 10
        titulo:
          type: string
          example: 'Dom Casmurro - edição revisada'
          nullable: true
        autor:
          type: string
          example: 'Machado de Assis'
          nullable: true
        ano:
          type: integer
          example: 1900
          nullable: true
        genero:
          type: string
          example: Romance
          nullable: true
        status:
          type: string
          example: lendo
          nullable: true
        avaliacao:
          type: integer
          example: 4
          nullable: true
        anotacoes:
          type: string
          example: 'Leitura em andamento'
          nullable: true
      type: object
    LivroDeleteRequest:
      required:
        - id_livro
      properties:
        id_livro:
          type: integer
          example: 10
      type: object
    LivroCreateResponse:
      properties:
        mensagem:
          type: string
          example: 'Livro criado com sucesso'
        id:
          type: integer
          example: 10
      type: object
    LivroUpdateResponse:
      properties:
        mensagem:
          type: string
          example: 'Livro atualizado com sucesso'
        detail:
          properties:
            livro:
              $ref: '#/components/schemas/LivroResource'
          type: object
      type: object
    LivroListResponse:
      properties:
        detail:
          properties:
            livros:
              type: array
              items:
                $ref: '#/components/schemas/LivroResource'
            paginacao:
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                total:
                  type: integer
                  example: 25
                total_pages:
                  type: integer
                  example: 3
              type: object
          type: object
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
    userUuid:
      type: apiKey
      description: 'UUID do usuário'
      name: X-USER-UUID
      in: header
...